refactor: colocate sprite avatar utility
